+++ AulaVisual/src/main/java/es/dws/aulavisual/controller/UserController.java
@@ -1,4 +1,4 @@
-package es.dws.aulavisual;
+package es.dws.aulavisual.controller;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletResponse;
@@ -6,12 +6,11 @@ import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import jakarta.annotation.PostConstruct;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import es.dws.aulavisual.users.UserService;
+import es.dws.aulavisual.service.UserService;
 import org.springframework.web.multipart.MultipartFile;
-import es.dws.aulavisual.users.User;
+import es.dws.aulavisual.model.User;
 
 import java.io.IOException;
 import java.util.Optional;
