+++ AulaVisual/src/main/java/es/dws/aulavisual/model/Module.java
@@ -1,6 +1,5 @@
-package es.dws.aulavisual.modules;
+package es.dws.aulavisual.model;
 
-import es.dws.aulavisual.courses.Course;
 import jakarta.persistence.*;
 import java.sql.Blob;
