+++ AulaVisual/src/main/java/es/dws/aulavisual/repository/ModuleRepository.java
@@ -1,6 +1,7 @@
-package es.dws.aulavisual.modules;
+package es.dws.aulavisual.repository;
 
-import es.dws.aulavisual.courses.Course;
+import es.dws.aulavisual.model.Course;
+import es.dws.aulavisual.model.Module;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Service;
