+++ AulaVisual/src/main/java/es/dws/aulavisual/repository/UserRepository.java
@@ -1,10 +1,9 @@
-package es.dws.aulavisual.users;
+package es.dws.aulavisual.repository;
 
-import es.dws.aulavisual.courses.Course;
+import es.dws.aulavisual.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
