+++ AulaVisual/src/main/java/es/dws/aulavisual/service/SubmissionService.java
@@ -1,7 +1,9 @@
-package es.dws.aulavisual.submissions;
+package es.dws.aulavisual.service;
 
-import es.dws.aulavisual.courses.Course;
-import es.dws.aulavisual.users.User;
+import es.dws.aulavisual.model.Course;
+import es.dws.aulavisual.model.Submission;
+import es.dws.aulavisual.repository.SubmissionRepository;
+import es.dws.aulavisual.model.User;
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
